move: constants를 utils 밖으로 이동
